requirements.txt
Django>=5.0,<6.0
PyPDF2>=3.0.0
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

